src/test/javascript/spec/app/entities/build-statistic/build-statistic-update.component.spec.ts
/*
 * Copyright (c) 2022 Eric Thiebaut-George.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */

import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { EnvKeeperTestModule } from '../../../test.module';
import { BuildStatisticUpdateComponent } from 'app/entities/build-statistic/build-statistic-update.component';
import { BuildStatisticService } from 'app/entities/build-statistic/build-statistic.service';
import { BuildStatistic } from 'app/shared/model/build-statistic.model';

describe('Component Tests', () => {
  describe('BuildStatistic Management Update Component', () => {
    let comp: BuildStatisticUpdateComponent;
    let fixture: ComponentFixture<BuildStatisticUpdateComponent>;
    let service: BuildStatisticService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [EnvKeeperTestModule],
        declarations: [BuildStatisticUpdateComponent],
        providers: [FormBuilder],
      })
        .overrideTemplate(BuildStatisticUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(BuildStatisticUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(BuildStatisticService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new BuildStatistic(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new BuildStatistic();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
